+++ tests/test_asend.py
@@ -211,3 +211,34 @@ async def test_edit_message_text_request_mocking(
         assert isinstance(response, tg_methods.EditMessageTextResponse)
         assert edit_message_text_response == json.loads(json_payload)
         assert edit_message_text_response == response.dict()
+
+
+@pytest.mark.anyio
+async def test_edit_message_reply_markup_request_mocking(
+    httpx_mock: pytest_httpx.HTTPXMock,
+    edit_message_reply_markup_response: dict[str, typing.Any],
+    keyboard: tg_types.InlineKeyboardMarkup,
+):
+    tg_types.Chat.update_forward_refs()
+    tg_types.Message.update_forward_refs()
+    httpx_mock.add_response(
+        url='https://api.telegram.org/bottoken/editmessagereplymarkup',
+        method='POST',
+        headers={
+            'content-type': 'application/json',
+            'accept': 'application/json',
+        },
+        json=edit_message_reply_markup_response,
+    )
+
+    async with tg_methods.AsyncTgClient.setup('token'):
+        tg_request = tg_methods.EditMessageReplyMarkupRequest(
+            chat_id=1234567890,
+            message_id=12345,
+            reply_markup=keyboard,
+        )
+        json_payload = await tg_request.apost_as_json('editmessagereplymarkup')
+        response = await tg_request.asend()
+        assert isinstance(response, tg_methods.EditMessageReplyMarkupResponse)
+        assert edit_message_reply_markup_response == json.loads(json_payload)
+        assert edit_message_reply_markup_response == response.dict()